Remove RoomController, RoomRepository, and RoomRequest. Change return type in Repository classes.
The Room-related files were deleted due to a change in the business requirements of the project. Room functionality is no longer needed, hence these files are removed. The return types of the find methods in some of the Repository classes were changed from List to Optional<List> to ensure the correct handling of the returned data. Additionally, new Custom Exception classes were created for better error handling. Lastly, changes were made in PaymentController.java to accurately check if the user has sufficient balance for a payment and to correctly convert the amount to Euro if required.